backend/requirements.txt
Flask
flask-cors
google-generativeai
requests
python-dotenv
vaderSentiment==3.3.2

